% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LargeDataExplorer-package.R
\docType{package}
\name{LargeDataExplorer-package}
\alias{LargeDataExplorer}
\alias{LargeDataExplorer-package}
\title{LargeDataExplorer: Powerful package to clean and re-format very large datasets after classifying its variables by their usefulness for machine learning}
\description{
This package can identify the irrelevant variables for machine learning and exclude them, while keeping and tranforming the useful ones. It also can explore a dataset, categorizing its variables as numeric, categorical (text), categorical (number), boolean, date, primary keys, keys, NA-variable, Univalue, plain text and variables with repeated information. It will also calculate the basic statistics for each variable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nietodaniel/LargeDataExplorer}
  \item Report bugs at \url{https://github.com/nietodaniel/LargeDataExplorer/issues}
}

}
\author{
\strong{Maintainer}: Daniel Nieto \email{nieto.daniel221@gmail.com} (\href{https://orcid.org/0000-0002-8609-3388}{ORCID})

}
\keyword{internal}
