% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{LDE.UsefulVars}
\alias{LDE.UsefulVars}
\title{Retrieving useful variables for analytics}
\usage{
LDE.UsefulVars(maxNARate = NULL, LDE.ExploreObject, ...)
}
\arguments{
\item{maxNARate}{numeric vector 0-1. Variables with a higher rate of NAs
will be excluded. Null to ignore}

\item{LDE.ExploreObject}{LDE.Explore() Object}

\item{...}{LDE.Explore() Objects}
}
\value{
A list containing the useful variables classified by its type with
their descriptive statistics and a vector containing the useful variable names
}
\description{
Classifies variables by its type and usefulness for analytics (E.g. Na-only,
1-value only, plain text types are not useful). It also includes descriptive
statistics for each included and excluded variable.
}
\examples{
df1 <- data.frame(secop1.full)
df2 <- data.frame(secop1.multas)
Explore.df1 <- LDE.Explore(df1)
Explore.df2 <- LDE.Explore(df2)

maxNARate <- 0.2

#Basic variable filtering of a single dataset exploration
Usefulness.1.df1 <- LDE.UsefulVars(NULL,Explore.df1)

#Basic variable filtering of a multiple datasets explortaions
Usefulness.1 <- LDE.UsefulVars(NULL,Explore.df1,Explore.df2)

#Variable filtering, including NA threshold, using multiple datasets explortaions
Usefulness.2 <- LDE.UsefulVars(maxNARate,Explore.df1,Explore.df2)

#df1: See if variables were included or excluded
#View(Usefulness.2$var.status$df1$included) #included vars
#View(Usefulness.2$var.status$df1$excluded) #excluded vars

#df1: See how the included variables were classified E.g.:
#View(Usefulness.2$var.classif$df1$included.vars$df.num) #numeric variables
#View(Usefulness.2$var.classif$df1$included.vars$df.bool) #boolean variables

#df1: See how the excluded variables were classified E.g.:
#View(Usefulness.2$var.classif$df1$removed.vars$not.useful) #excluded by type
#View(Usefulness.2$var.classif$df1$removed.vars$filtered.byNAs) #excluded by NA rate
#View(Usefulness.2$var.classif$df1$removed.vars$not.useful$df.NA) #excluded by type, empty
#View(Usefulness.2$var.classif$df1$removed.vars$filtered.byNAs$df.num) #numeric, excluded by NA rate

#df1: See statistics of variables by exclusion reason
#View(Usefulness.2$statistics$df1$useful.vars) #included
#View(Usefulness.2$statistics$df1$filteredbyNAs.vars) #excluded by NA rate
#View(Usefulness.2$statistics$df1$unuseful.vars) #excluded by type

#df1: See statistics of variables by exclusion reason and type E.g.:
#View(Usefulness.2$statistics$df1$useful.vars$df.levels) #included, type level
#View(Usefulness.2$statistics$df1$filteredbyNAs.vars$df.num) #numeric, excluded by NA rate
#View(Usefulness.2$statistics$df1$unuseful.vars$df.NA) #excluded, type empty
}
\author{
Daniel Nieto
}
