% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secop1.full}
\alias{secop1.full}
\title{State contracts dataset. Each row is a state procurement process.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 200 rows and 72 columns.
}
\usage{
secop1.full
}
\description{
Modified and shortened datased based on "SECOP I - 2020: Data set of
the processes published in SECOP I during the year 2020", a dataset
published under the Colombian government's project "Colombia Compra
Eficiente" which seeks to reduce corruption by publicly publishing
the state procurement processes. https://www.colombiacompra.gov.co/
The original dataset was published under the Open Contracting Data
Standard License (ODCS), which means it is open for any use.
The authors of this package have no role in the Colombian government
}
\keyword{datasets}
