% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secop1.multas}
\alias{secop1.multas}
\title{Fines and sanctions dataset. Each row is a fine.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 699 rows and 12 columns.
}
\usage{
secop1.multas
}
\description{
Modified and shortened datased based on "Multas y Sanciones SECOP I:
Dataset of the fines and sanctions generated in the SECOP I platform",
a dataset published under the Colombian government's project "Colombia
Compra Eficiente" which seeks to reduce corruption by publicly publishing
the state procurement processes. https://www.colombiacompra.gov.co/
The original dataset was published under the Open Contracting Data
Standard License (ODCS), which means it is open for any use.
The authors of this package have no role in the Colombian government
}
\keyword{datasets}
