% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LargeDataExplorer-package.R
\docType{package}
\name{LargeDataExplorer-package}
\alias{LargeDataExplorer}
\alias{LargeDataExplorer-package}
\title{LargeDataExplorer: Fast and powerful package for preliminary exploration of large datasets}
\description{
This package explores a dataset, categorizing its variables as numeric, categorical (text), levels (categorical numbers), boolean, plain text, Primary Key (A different value for each row), NA (Empty variable)or Univalue (Only one value in the whole variable). It will also calculate basic statistics for each variable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nietodaniel/LargeDataExplorer}
  \item Report bugs at \url{https://github.com/nietodaniel/LargeDataExplorer/issues}
}

}
\author{
\strong{Maintainer}: Daniel Nieto \email{nieto.daniel221@gmail.com} (\href{https://orcid.org/0000-0002-8609-3388}{ORCID})

}
\keyword{internal}
