% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{LDE.AutoProcess}
\alias{LDE.AutoProcess}
\title{data.frame automatic exploration, filtering and format transformation}
\usage{
LDE.AutoProcess(dat, maxNARate = NULL)
}
\arguments{
\item{dat}{data.frame}

\item{maxNARate}{numeric vector 0-1. Null to ignore}
}
\value{
The filtered dataset, with re-formatted variables and all the
process information including descriptive statistics
}
\description{
Automatically removes unuseful variables (Empty, plain text, one-value,
with repeated information) and variables with excess NAs.
}
\examples{
dat <- data.frame(iris)
maxNARate = 0.2
Exploration.dat <- LDE.AutoProcess(dat,maxNARate)

#View all the process information
#Includes descriptive statistics for each variable
str(Exploration.dat$process.info)

#Obtains clean dataset
dat.clean<-Exploration.dat$df.filtered
str(dat.clean)
}
\author{
Daniel Nieto
}
