% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{LDE.Explore}
\alias{LDE.Explore}
\title{data.frame preliminar exploration}
\usage{
LDE.Explore(dat)
}
\arguments{
\item{dat}{data.frame}
}
\value{
A list containing the variables classified by its type and their
descriptive statistics
}
\description{
classifies a data.frame variables by type (numeric, categorical, boolean,
text, etc.) and performs descriptive statistics for each (mean, min, max,
sd, skewness, etc).
}
\examples{
df <- data.frame(iris)
Explore.df <- LDE.Explore(df)

#View which variables are of each type and its characteristics
str(Explore.df$df.text) #Variables of text type
str(Explore.df$df.num) #Variables of numeric type
str(Explore.df$df.NA) #Empty variables
str(Explore.df$df.univalue) #Variables with only one value
str(Explore.df$df.category) #Variables of categorical (text) type
str(Explore.df$df.levels) #Variables of categorical (numeric) type
str(Explore.df$df.bool) #Variables of bool type
str(Explore.df$df.primarykeys) #Variables which could be primary keys
str(Explore.df$df.repeatedVars) #Variables with repeated information
}
\author{
Daniel Nieto
}
