% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exports.R
\name{LDE.UsefulVars}
\alias{LDE.UsefulVars}
\title{useful variables selection}
\usage{
LDE.UsefulVars(maxNARate = NULL, LDE.ExploreObject, ...)
}
\arguments{
\item{maxNARate}{numeric vector 0-1. Null to ignore}

\item{LDE.ExploreObject}{LDE.Explore() Object}

\item{...}{LDE.Explore() Objects}
}
\value{
A list containing the useful variables classified by its type with
their descriptive statistics and a vector containing the useful variable names
}
\description{
After performing LDE.Explore(), the output can be filtered to
obtain only the useful varibles and the variables names in a vector. A
max rate of NAs (0-1) can be set to filther further
}
\examples{
df.iris <- data.frame(iris)
df.cars <- data.frame(cars)
Explore.iris <- LDE.Explore(df.iris)
Explore.cars <- LDE.Explore(df.cars)
maxNARate <- 0.2 #All variables with more than 20\% of NAs will be excluded
usefulVars <- LDE.UsefulVars(maxNARate,Explore.iris,Explore.cars)

#Useful Variables --> usefulVars$varnames
str(usefulVars$varnames$df.iris) #Useful varnames of Iris dataset
str(usefulVars$varnames$df.cars) #Useful varnames of Cars dataset

#View useful variables of each dataset and its characteristics
str(usefulVars$df.num) #Variables of numeric type
str(usefulVars$df.category) #Variables of categorical (text) type
str(usefulVars$df.bool) #Variables of bool type
str(usefulVars$df.levels) #Variables of categorical (numeric) type
str(usefulVars$df.primarykeys) #Variables which could be primary keys
str(usefulVars$df.keys) #Variables which could be keys
}
\author{
Daniel Nieto
}
